// src/app/types.ts
export interface CourseType {
  id: number;
  title: string;
  description: string;
  lectures: number;
  duration: string;
}

export interface Course {
  id: number;
  title: string;
  parag_one: string;
  parag_two_solid: string;
  parag_three_bold: string;
  list: string[];
  parag_four_solid: string;
  parag_four_bold: string;
  partone_hone: string;
  partone_p: string;
  partone_img: string;
  parttwo_hone: string;
  parttwo_p: string;
  partthree_title: string;
  partthree_ul: string[];
  partfour_img: string;
  partfive_pone: string;
  partfive_ptwo: string;
  partorange: string;
}